% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_file.R
\name{add_1_to_x}
\alias{add_1_to_x}
\title{Function for writing a tibble as a table the database
This function will: i) connect to database, ii) write table; append if desired, iii) exit connection}
\usage{
add_1_to_x(x)
}
\arguments{
\item{x}{column use to compare the old and new data; only new data unique in this column will be added (eg. use date or a key/id)}
}
\description{
Function for writing a tibble as a table the database
This function will: i) connect to database, ii) write table; append if desired, iii) exit connection
}
\examples{
New example, test online
}
\keyword{database}
