% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_file.R
\name{queryDatabaseUser}
\alias{queryDatabaseUser}
\title{Function to query the postgres database from R
Returns the result of the query in dataframe format}
\usage{
queryDatabaseUser(query)
}
\arguments{
\item{query}{needs to be in postgres code}
}
\description{
Function to query the postgres database from R
Returns the result of the query in dataframe format
}
\examples{
queryDatabase("SELECT datetime, precip_tot FROM z_micromet WHERE datetime BETWEEN '01/01/2009 00:00' AND '01/01/2010 00:00'")
}
