% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_file.R
\name{showColumns}
\alias{showColumns}
\title{Function to show columns in a table}
\usage{
showColumns(tablename)
}
\arguments{
\item{tablename}{is the name of the table for which columns should be shown}
}
\description{
Function to show columns in a table
}
