% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_file.R
\name{showRange}
\alias{showRange}
\title{Function to show the class and range of data in a column}
\usage{
showRange(tablename, colnames)
}
\arguments{
\item{tablename}{is the name of the table}

\item{colname}{is the name(s) of the columns}
}
\description{
Function to show the class and range of data in a column
}
