% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_file.R
\name{showSize}
\alias{showSize}
\title{Function to get the size of a table}
\usage{
showSize(tablename)
}
\arguments{
\item{tablename}{is the name of the table}
}
\description{
Function to get the size of a table
}
